requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
httpx>=0.24.0

# Docker support
docker>=6.1.0

# Database & Storage
supabase>=1.0.0
asyncpg>=0.27.0
redis>=5.0.0
aioredis>=2.0.0
python-magic>=0.4.27  # For MIME type detection

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
python-magic-bin>=0.4.14; platform_system == "Windows"  # For Windows support
